packages/vision-analysis-mcp/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0

# OCR Only (lightweight)
paddleocr>=2.7.0
paddlepaddle>=2.6.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Claude API
anthropic>=0.50.0

# Utilities
python-dotenv>=1.0.0
